move former LabelBuilder into text style (specific to text labels now, point style has its own version)
